+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 2.8, 2024-12-03
 ---------------
+- add program version update check
 - fix saving slots that were initially empty then populated then emptied again
 - fix tracking changes over undo-redo after saving
 - fix error after confirming to save changes when closing file